requirements.txt
numpy
matplotlib
pandas
scipy
torch
prettytable
pillow
lpips
